requirements.txt
python-dotenv==0.10.3
elasticsearch==7.0.2
pytz==2019.1
Flask==1.0.2
Flask-Cors==3.0.8
prometheus_client==0.7.1
requests==2.22.0
